refactor(review): bound completion contract modules
