+++ crates/cli-sub-agent/src/review_convergence/completion.rs
@@ -1,276 +1,14 @@
 use std::collections::HashSet;
-use std::error::Error;
-use std::fmt;
 use std::future::Future;
 use std::pin::Pin;
 
-use csa_session::convergence::{
-    AdmittedModelIdentity, ArtifactEvidenceRef, CampaignId, CandidateId, EpochId, EpochRecord,
-    RepairBatchId, RootClusterId,
+use csa_session::convergence::{CampaignId, CandidateId, EpochId, EpochRecord, RootClusterId};
+
+pub(crate) use super::completion_types::{
+    AuthorizedRepairBatch, CompletionAction, CompletionBudget, CompletionError, CompletionEvent,
+    CompletionOutcome, CompletionPhase, CompletionPortError, CompletionState, CompletionTransition,
 };
-
-use super::discovery_contract::{
-    CampaignSelection, CleanRoomReviewOutput, DiscoveryFocus, TargetedDiscoveryFocus,
-};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) struct CompletionBudget {
-    max_cycles: u32,
-    max_provider_actions: u32,
-}
-
-impl CompletionBudget {
-    pub(crate) fn new(max_cycles: u32, max_provider_actions: u32) -> Result<Self, CompletionError> {
-        if max_cycles == 0
-            || max_provider_actions == 0
-            || max_cycles > 10_000
-            || max_provider_actions > 1_000
-        {
-            return Err(CompletionError::InvalidBudget);
-        }
-        Ok(Self {
-            max_cycles,
-            max_provider_actions,
-        })
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum CompletionPhase {
-    Start,
-    Discover,
-    VerifyAndCluster,
-    RunAuthorizedRepairs,
-    RunFinalGates,
-    RunFreshCleanRoom,
-    PublishFinalPair,
-    Attested,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct AuthorizedRepairBatch {
-    root_cluster_id: RootClusterId,
-    repair_batch_id: RepairBatchId,
-}
-
-impl AuthorizedRepairBatch {
-    pub(crate) fn new(root_cluster_id: RootClusterId, repair_batch_id: RepairBatchId) -> Self {
-        Self {
-            root_cluster_id,
-            repair_batch_id,
-        }
-    }
-
-    pub(crate) fn root_cluster_id(&self) -> &RootClusterId {
-        &self.root_cluster_id
-    }
-
-    pub(crate) fn repair_batch_id(&self) -> &RepairBatchId {
-        &self.repair_batch_id
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct CompletionState {
-    phase: CompletionPhase,
-    budget: CompletionBudget,
-    cycles: u32,
-    provider_actions: u32,
-    campaign_id: Option<CampaignId>,
-    epoch: Option<EpochRecord>,
-    discovery_focus: Option<DiscoveryFocus>,
-    campaign_selection: Option<CampaignSelection>,
-    pending_candidates: Vec<CandidateId>,
-    repair_batches: Vec<AuthorizedRepairBatch>,
-    gate_artifact: Option<ArtifactEvidenceRef>,
-    clean_room: Option<CleanRoomReviewOutput>,
-}
-
-impl CompletionState {
-    pub(crate) fn new(budget: CompletionBudget) -> Self {
-        Self {
-            phase: CompletionPhase::Start,
-            budget,
-            cycles: 0,
-            provider_actions: 0,
-            campaign_id: None,
-            epoch: None,
-            discovery_focus: None,
-            campaign_selection: None,
-            pending_candidates: Vec::new(),
-            repair_batches: Vec::new(),
-            gate_artifact: None,
-            clean_room: None,
-        }
-    }
-
-    pub(crate) fn phase(&self) -> CompletionPhase {
-        self.phase
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum CompletionAction {
-    Discover {
-        focus: DiscoveryFocus,
-        selection: CampaignSelection,
-        epoch: EpochRecord,
-    },
-    VerifyAndCluster {
-        campaign_id: CampaignId,
-        epoch: EpochRecord,
-        candidates: Vec<CandidateId>,
-    },
-    RunAuthorizedRepairs {
-        campaign_id: CampaignId,
-        epoch: EpochRecord,
-        batches: Vec<AuthorizedRepairBatch>,
-    },
-    RunFinalGates {
-        campaign_id: CampaignId,
-        epoch: EpochRecord,
-    },
-    RunFreshCleanRoom {
-        campaign_id: CampaignId,
-        epoch: EpochRecord,
-        gate_artifact: ArtifactEvidenceRef,
-    },
-    PublishFinalPair {
-        campaign_id: CampaignId,
-        epoch: EpochRecord,
-        gate_artifact: ArtifactEvidenceRef,
-        clean_room: Box<CleanRoomReviewOutput>,
-    },
-}
-
-impl CompletionAction {
-    fn consumes_provider_action(&self) -> bool {
-        matches!(
-            self,
-            Self::Discover { .. }
-                | Self::VerifyAndCluster { .. }
-                | Self::RunAuthorizedRepairs { .. }
-                | Self::RunFreshCleanRoom { .. }
-        )
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum CompletionEvent {
-    Started {
-        epoch: EpochRecord,
-        selection: CampaignSelection,
-    },
-    DiscoveryCompleted {
-        focus: DiscoveryFocus,
-        selection: CampaignSelection,
-        campaign_id: CampaignId,
-        epoch: EpochRecord,
-        candidates: Vec<CandidateId>,
-    },
-    ClustersReady {
-        campaign_id: CampaignId,
-        epoch_id: EpochId,
-        verified_candidates: Vec<CandidateId>,
-        root_clusters: Vec<RootClusterId>,
-        repair_batches: Vec<AuthorizedRepairBatch>,
-    },
-    RepairsCompleted {
-        campaign_id: CampaignId,
-        previous_epoch_id: EpochId,
-        completed_batches: Vec<RepairBatchId>,
-        new_epoch: EpochRecord,
-    },
-    FinalGatesPassed {
-        campaign_id: CampaignId,
-        epoch_id: EpochId,
-        artifact: ArtifactEvidenceRef,
-    },
-    CleanRoomCompleted {
-        campaign_id: CampaignId,
-        epoch_id: EpochId,
-        output: CleanRoomReviewOutput,
-    },
-    FinalPairPublished {
-        campaign_id: CampaignId,
-        epoch_id: EpochId,
-        gate_artifact: ArtifactEvidenceRef,
-        review_artifact: ArtifactEvidenceRef,
-        model_identity: AdmittedModelIdentity,
-    },
-    DriftDetected,
-    CleanupUncertain,
-    ProviderUnavailable,
-    IncompleteProviderOutput,
-    MaxRoundsReached,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum CompletionOutcome {
-    Attested {
-        campaign_id: CampaignId,
-        epoch: EpochRecord,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct CompletionTransition {
-    pub(crate) state: CompletionState,
-    pub(crate) action: Option<CompletionAction>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum CompletionError {
-    InvalidBudget,
-    InvalidTransition(&'static str),
-    IdentityMismatch,
-    DuplicateIdentity,
-    CardinalityMismatch,
-    EpochDidNotChange,
-    BudgetExhausted,
-    BlockedCleanRoom,
-    DriftDetected,
-    CleanupUncertain,
-    ProviderUnavailable,
-    IncompleteProviderOutput,
-    MaxRoundsReached,
-    Port(String),
-}
-
-impl fmt::Display for CompletionError {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::InvalidBudget => formatter.write_str("completion budget is invalid"),
-            Self::InvalidTransition(message) => {
-                write!(formatter, "invalid completion transition: {message}")
-            }
-            Self::IdentityMismatch => formatter.write_str("completion event identity mismatch"),
-            Self::DuplicateIdentity => {
-                formatter.write_str("completion event contains a duplicate identity")
-            }
-            Self::CardinalityMismatch => {
-                formatter.write_str("root cluster and repair batch cardinality mismatch")
-            }
-            Self::EpochDidNotChange => {
-                formatter.write_str("authorized repair did not create a changed HEAD epoch")
-            }
-            Self::BudgetExhausted => formatter.write_str("completion budget exhausted"),
-            Self::BlockedCleanRoom => {
-                formatter.write_str("clean-room questions or unchecked items block completion")
-            }
-            Self::DriftDetected => formatter.write_str("completion input drift detected"),
-            Self::CleanupUncertain => formatter.write_str("completion cleanup is uncertain"),
-            Self::ProviderUnavailable => formatter.write_str("completion provider unavailable"),
-            Self::IncompleteProviderOutput => {
-                formatter.write_str("completion provider output is incomplete")
-            }
-            Self::MaxRoundsReached => formatter.write_str("completion maximum rounds reached"),
-            Self::Port(message) => write!(formatter, "completion port failed: {message}"),
-        }
-    }
-}
-
-impl Error for CompletionError {}
+use super::discovery_contract::{CampaignSelection, DiscoveryFocus, TargetedDiscoveryFocus};
 
 pub(crate) fn reduce_completion(
     state: &CompletionState,
@@ -613,23 +351,6 @@ fn validate_cluster_batches(
     }
     Ok(())
 }
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct CompletionPortError(String);
-
-impl CompletionPortError {
-    pub(crate) fn new(message: impl Into<String>) -> Self {
-        Self(message.into())
-    }
-}
-
-impl fmt::Display for CompletionPortError {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter.write_str(&self.0)
-    }
-}
-
-impl Error for CompletionPortError {}
 
 pub(crate) trait CompletionPorts {
     fn execute<'a>(
